Add files via upload
Commit Message: Initial commit of the time series data scraper

Description: This commit includes the initial version of the Python script for scraping time series data from the specified API endpoint. The script makes HTTP requests, parses JSON responses, organizes the data, and exports it to a CSV file. This is the starting point of the project.